+++ rstview/templatetags/rstview_tags.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Parser template tags 
+Parser template tags
 """
 from django import template
 from django.utils.safestring import mark_safe
@@ -9,10 +9,23 @@ from rstview.parser import SourceParser
 
 register = template.Library()
 
-def source_render(source, setting_key="default"):
+
+@register.simple_tag
+def rst_render(source, *args, **kwargs):
     """
     Return the parser result from the given string source and settings
+
+    {% rst_render SOURCE_STRING [config='default'] [body_only=True] [silent=False] %}
     """
-    return mark_safe( SourceParser(source, setting_key=setting_key) )
-source_render.is_safe = True
-register.filter(source_render)
+    config_name = kwargs.get('config', 'default')
+    body_only = kwargs.get('body_only', True)
+    silent = kwargs.get('silent', False)
+
+    return mark_safe(
+        SourceParser(
+            source,
+            setting_key=config_name,
+            body_only=body_only,
+            silent=silent,
+        )
+    )