+++ HISTORY.rst
@@ -15,6 +15,7 @@ Version 0.3.0 - Unreleased
   * ``rstview.parser.map_parsing_errors`` has moved to ``rstview.reporter.format_parsing_errors``;
   * Improved view ``rstview.views.RSTFileView`` to be more flexible;
   * Default shipped view template now inherits from ``skeleton.html`` instead of ``base.html``;
+  * Template filter ``foo`` has been dropped in profit of template tag ``bar`` which has more options;
 
 Version 0.2.1 - 2014/08/17
 --------------------------